Process and Queue Header
This header file provides a process class and a processQueue class. This gives us the outline for a process as well as built-in queues. We can figure out more details for the logic of the queues later -- will most likely be done in each of the algo files -- ie, we can handle the qs depending on what algo the user wants (defined by the switch cases in the driver prog).